Remove OTP cache (again)
The sqlite cache wouldn't work in an HA system as one can never
tell which controller is going to be requested. This would require
a shared or replicating database backend which is simply overkill
for this.

Instead let the last request win. With each request for metadata
generate a new OTP and update the value in the host record in
IPA.