Let the last metadata fetch win in setting the OTP
It appears that the metadata can be retrieved multiple times
when standing up a new instance. This is problematic for OTP
because we only have the clear value during the add phase so
if the host already exists we blow up.

I had been using a cache to store the value based on instance_id
but it wouldn't scale or work with HA without using a real
database backend. So instead the last update wins.