Add nova client for retrieving instance information
This adds a NovaClient class and a method for obtaining
the instance object from nova using the instance_id
provided by nova.

Right this this is only used to determine if the instance
exists in nova at all. Relying on an internal nova structure
is generally frowned upon by the nova core team.