Require a valid image id to enroll a host
I was previously ignoring an invalid or inaccessible image but
I think this should instead be a hard failure. How can a user
request an image create in nova with an image that can't be read?

I also noticed that it was possible for any test looking for a
raised error to fail if that exception wasn't raised so
assert(False) in that case.