Fix test case where IPA is not locally configured
These tests are expected to run outside the context of an
available IPA server but there are still two conditions.

The first is no IPA is configured so that will be detected
and no OTP returned. In this case no OTP is returned so there
is nothing to test against.

The second is IPA is configured but there is no keytab so
no host can be added. In this case the OTP is still returned
so it can be checked at least that it matches the expected
pattern.

Change-Id: I01672ecdef00728efb76339506946047651a4c13