Add py35 to environments in tox.ini
This will effectively run the tests with py35. Note that you might hit
issues when running this if you have previously ran tox with the py27
environment. This might be present with the error message:

    db type could not be determined

If this happens, it's a matter of removing the following file:
.testrepository/times.dbm

this issue is documented here:
https://bugs.launchpad.net/testrepository/+bug/1229445

Change-Id: I867457a30e07fad1a64111b5bf6479529e6fc19b