Extend the novajoin-ipa-setup script to precreate IPA entries
For the case of Triple-O we don't want to pass IPA admin
credentials to the undercloud so instead pre-create the IPA
entries for undercloud and pass in an OTP that can be used to
enroll it using ipa-client-install.

Another feature of --precreate is that it doesn't require the
machine to be enrolled as an IPA client. The required options must
be provided on the command-line.

Change-Id: Ia69b5b4fbc275c04f5e07e9d2ef62e3547725ac8