Handle compact service representation
this adds a 'compact_services' option for the metadata that we take from
nova which gets novajoin to parse a more compact representation of
services for the given node. This was done to address the fact that the
nova metadata service has a limit of 255 characters for the value of a
given key, and this was problematic for dealing with sub-hosts and their
respective services.

Change-Id: I440a70567b396a62b56052b6501875ac537762ca