Handle services with unique keys from metadata
This uses the same mechanism for creating services from complete
service principals, however, they are assumed to come from an individual
and unique key with a defined prefix. In this case, the prefix is
assumed to be 'managed_service_'. We then iterate these keys from the
metadata and form the list of service principals we need to create and
manage. Whatever comes after the prefix is discarded, as all we really
need is for it to be unique.

Change-Id: Ic7b2b060fb597f82beebdc8ed1279a866dfe6e36