Don't use a service credential anymore, use token nova passes in
Since nova is now authenticating its requests to the dynamic
metadata services use the token that nova passes in to verify
the image and instance information.

This moves the auth configuration into the [keystone_authtoken]
section. This will be used to validate the incoming token and
provide the auth_url needed to re-use the token for other requests.

Change-Id: Ia2ef4dec983f4732bbc0e10870cd2873a7302228