Fix IPA v4.5.0 import issue with kinit_keytab
The import was wrapped in a try/except to continue on import
failures so the unit tests will work which masked the fact that
kinit_keytab had changed locations. This meant that no credentials
were obtained and api.finalize() never called so subequent attempts
to use api.env.* would fail.

Change-Id: I6016892630510b816721cea5b20c8d6e7f8d34fa