Add basic service and host caches
This adds two caches: one for hosts and another one for services. The
service cache also contains which hosts are managing the service.

This was done in order to reduce the calls to FreeIPA and to try to make
novajoin slightly more efficient.

Note that this was only added to the "add" functions, and the delete
functions merely update the cache. This is because checking for hosts
managing a group would require the cache to be consistent between all
the processes (and novajoin could be ran in several), and for this the
best thing would be to use a distributed cache. Being this the first
attempt, we leave this functionality out of the scope for this patch.

Change-Id: Id107000b3a667f5724331e281912560cff6f92f0