Add a retry loop when fetching vendor_data from novajoin
In [1] we added a random sleep in the script called by cloud-init
to fetch vendor_data, to lower chances that many nodes connect to
novajoin concurrently. Still, curl failures still happen from
time to time.

Add a retry loop around the call to curl to further increase the
chances of success, as this seems to give good results in practice
when deploying overclouds with tripleo.

[1] 188f09ef22a233323464fcb0a8c6942d6981032a

Closes-Bug: #1766824

Change-Id: I728c2cb0c8a7433b68dd7de2de242e922974d713