Fix - Invalid ipaotp returned if host in cache
Change: Id107000b3a667f5724331e281912560cff6f92f0 implemented
caching in the IPAClient. We need to store the OTP in the cache
and return the cached OTP, not the one generated on the join
request in case there is a cache hit, since we do not update
the OTP in FreeIPA when the host is in the cache.

Change-Id: Ic19ee7c2228d275397bc4be04432126fd2f228ec
Closes-Bug: 1796415